Added HCHO weekend/weekday solver
This one requires the VOCR to be the same on weekends and weekdays
Which required splitting out the HCHO steady state equation to
its own function (`private/hcho_steady_state.m`). Also updated the
original HCHO constraint to make the cost from differing HCHO
concentrations actually be the same order of magnitude as lifetime.